Added support for theme inheritance.
Themes are now specified as an array rather than a single element. Files are
looked for in templates (and static directories) in the order they are listed.
This way, it is possible to define a new theme that makes only minor modifications of the default theme without copying all the files.